+++ JavaScriptEngineSwitcher.Core/JsEngineNotFoundException.cs
@@ -8,7 +8,7 @@
 	public sealed class JsEngineNotFoundException : Exception
 	{
 		/// <summary>
-		/// Initializes a new instance of the JavaScriptEngineSwitcher.Core.JsEngineNotFoundException class 
+		/// Initializes a new instance of the <see cref="JsEngineNotFoundException"/> class 
 		/// with a specified error message
 		/// </summary>
 		/// <param name="message">The message that describes the error</param>
@@ -17,7 +17,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the JavaScriptEngineSwitcher.Core.JsEngineNotFoundException class 
+		/// Initializes a new instance of the <see cref="JsEngineNotFoundException"/> class 
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>