+++ JavaScriptEngineSwitcher.Core/Utilities/Utils.cs
@@ -120,5 +120,31 @@
 
 			return (T)instance;
 		}
+
+		/// <summary>
+		/// Converts value of source enumeration type to value of destination enumeration type
+		/// </summary>
+		/// <typeparam name="TSource">Source enumeration type</typeparam>
+		/// <typeparam name="TDest">Destination enumeration type</typeparam>
+		/// <param name="value">Value of source enumeration type</param>
+		/// <returns>Value of destination enumeration type</returns>
+		public static TDest GetEnumFromOtherEnum<TSource, TDest>(TSource value)
+		{
+			string name = value.ToString();
+			var destEnumValues = (TDest[])Enum.GetValues(typeof(TDest));
+
+			foreach (var destEnum in destEnumValues)
+			{
+				if (string.Equals(destEnum.ToString(), name, StringComparison.OrdinalIgnoreCase))
+				{
+					return destEnum;
+				}
+			}
+
+			throw new InvalidCastException(
+				string.Format(Strings.Common_EnumValueConversionFailed,
+					name, typeof(TSource), typeof(TDest))
+			);
+		}
 	}
 }